refactor(mcp): split test files to match source module structure
